+++ src/kernel_sdk/CHANGES.md
@@ -32,6 +32,7 @@
 - Remove redundant `testing` feature flag from `host` crate.
 - Disable `testing` as default features of the main `sdk` crate.
 - Change `Runtime::store_value_size` to return `PathNotFound` when the input is the path of a directory.
+- Add support for `MichelsonTimestamp` encoding and decoding.
 
 ### Installer client/kernel